# Feature/flatten struct, but with BFS