--- a/harley/transformations.py
+++ b/harley/transformations.py
@@ -3,7 +3,7 @@ from .harley import columns_to_snake_case
 from typing import Union, List, Dict
 from polars import col, Struct, Expr, LazyFrame
 
-import logging
+import warnings
 
 
 def snake_case_column_names(df: PolarsFrame) -> PolarsFrame:
@@ -13,6 +13,7 @@ def snake_case_column_names(df: PolarsFrame) -> PolarsFrame:
 
 class ColumnNameRepeatedError(ValueError):
     """raised when a column name would be repeated after flatten_struct"""
+
 
 def flatten_struct(
     df: PolarsFrame,
@@ -31,32 +32,38 @@ def flatten_struct(
     if limit is not None and limit < 0:
         raise ValueError("limit must be a positive integer or None")
     if limit == 0:
-        logging.warning("limit of 0 will result in no transformations")
+        warnings.warn("limit of 0 will result in no transformations")
         return df
-    ldf = df.lazy() # noop if df is LazyFrame
+    ldf = df.lazy()  # noop if df is LazyFrame
     all_column_names = ldf.collect_schema().names()
     if any(separator in (witness := column) for column in all_column_names):
-        logging.warning(
-            f"separator \"{separator}\" found in column names, e.g. \"{witness}\". "
-            "This might columns to be repeated this flatten_struct function to error"
+        warnings.warn(
+            f'separator "{separator}" found in column names, e.g. "{witness}". '
+            "If columns would be repeated, this function will error"
         )
     non_struct_columns = list(set(ldf.collect_schema().names()) - set(struct_columns))
     struct_schema = ldf.select(*struct_columns).collect_schema()
     col_dtype_expr_names = [(struct_schema[c], col(c), c) for c in struct_columns]
     result_names: Dict[str, Expr] = {}
     level = 0
-    while (limit is None and col_dtype_expr_names) or (limit is not None and level < limit):
+    while (limit is None and col_dtype_expr_names) or (
+        limit is not None and level < limit
+    ):
         level += 1
         new_col_dtype_exprs = []
         for dtype, col_expr, name in col_dtype_expr_names:
             if not isinstance(dtype, Struct):
-                if drop_original_struct:
-                    if name in result_names:
-                        raise ColumnNameRepeatedError(
-                            f"Column name {name} would be repeated after flatten_struct"
-                        )
-                    result_names[name] = col_expr
+                if name in result_names:
+                    raise ColumnNameRepeatedError(
+                        f"Column name {name} would be created at least twice after flatten_struct"
+                    )
+                result_names[name] = col_expr
                 continue
+            if any(separator in (witness := field.name) for field in dtype.fields):
+                warnings.warn(
+                    f'separator "{separator}" found in field names, e.g. "{witness}" in {name}. '
+                    "If columns would be repeated, this function will error"
+                )
             new_col_dtype_exprs += [
                 (
                     field.dtype,
@@ -73,13 +80,17 @@ def flatten_struct(
                     for field in dtype.fields
                 )
         col_dtype_expr_names = new_col_dtype_exprs
-    if level == limit and col_dtype_expr_names:
+    if drop_original_struct and level == limit and col_dtype_expr_names:
         for _, col_expr, name in col_dtype_expr_names:
             result_names[name] = col_expr
-    if result_names and drop_original_struct:
+    if any((witness := column) in non_struct_columns for column in result_names):
+        raise ColumnNameRepeatedError(
+            f"Column name {witness} would be created after flatten_struct, but it's already a non-struct column"
+        )
+    if drop_original_struct:
         ldf = ldf.select(
-            [col(c) for c in non_struct_columns] +
-            [col_expr.alias(name) for name, col_expr in result_names.items()]
+            [col(c) for c in non_struct_columns]
+            + [col_expr.alias(name) for name, col_expr in result_names.items()]
         )
 
     if isinstance(df, LazyFrame):