--- a/requirements.txt
+++ b/requirements.txt
@@ -3,3 +3,4 @@ maturin
 ruff
 pytest
 mypy
+pytest-timeout