- refactor flatten_struct to perform the recursion using Breadth First Search instead of recursive calls
- add errors if it catches it would create the same column multiple times (e.g. `"a:b" : "c"` vs `"a" : "b" : "c"`), or create a column that's already in the dataframe itself. This is usualy because one or more of the fields / columns has the separator in its name.
- update tests
  - add tests for errors
  - also added pytest-timeout to requirements, though it's just dev-only. Useful for when I was refactoring and accidentally introduced infinite loops. Now it times out if the test doesn't finish in a fixed time (e.g. 1 second)

Using this benchmark, I found it could be twice as fast! (It was already very fast in the first place. Also 1000 runs might be a bit small.)
```text
# OLD
Generating dataframe
Done
Testing flatten_struct with recursive=True
for 1000000 rows for 1000 runs
It took 0.001018240371000502s on average
Testing flatten_struct with recursive=True, drop_original_struct=False
for 1000000 rows for 1000 runs
It took 0.0009654799259988067s on average
# NEW
Generating dataframe
Done
Testing flatten_struct with recursive=True
for 1000000 rows for 1000 runs
It took 0.0004731835689999571s on average
Testing flatten_struct with recursive=True, drop_original_struct=False
for 1000000 rows for 1000 runs
It took 0.0006911943859995518s on average
```
<details>
<summary>Benchmarking code</summary>

```python
import timeit
import polars as pl
from harley import flatten_struct

DATAFRAME_LENGTH = 1_000_000

def gen_dataframe(n: int) -> pl.DataFrame:
    print("Generating dataframe")
    rows = [
        {
            "a": {"b": {"c": x, "d": 2 * x}, "e": {"f": 3 * x, "g": 4 * x}},
            "p": {"q": True, "r": "bar", "s": 5.0 * x},
            "t": {"u": {"v": {"w": f"{x}"}}},
        }
        for x in range(n)
    ]
    df = pl.DataFrame(rows)
    print("Done")
    return df

df = gen_dataframe(DATAFRAME_LENGTH)
setup = "from __main__ import df, flatten_struct"
stmt = "res_df = flatten_struct(df, ['a', 'p', 't'], recursive=True)"

REPEAT_TIMES = 1000
print("Testing flatten_struct with recursive=True")
print(f"for {DATAFRAME_LENGTH} rows for {REPEAT_TIMES} runs")
t = timeit.timeit(setup=setup, stmt=stmt, number=REPEAT_TIMES)
print(f"It took {t/REPEAT_TIMES}s on average")

stmt = "res_df = flatten_struct(df, ['a', 'p', 't'], recursive=True, drop_original_struct=False)"
print("Testing flatten_struct with recursive=True, drop_original_struct=False")
print(f"for {DATAFRAME_LENGTH} rows for {REPEAT_TIMES} runs")
t = timeit.timeit(setup=setup, stmt=stmt, number=REPEAT_TIMES)
print(f"It took {t/REPEAT_TIMES}s on average")
```
</details>